refactor: TIO-172: move component into folder
