+++ lib/pages/metronome/setting_bpm.dart
@@ -9,7 +9,7 @@ import 'package:tiomusic/pages/parent_tool/parent_setting_page.dart';
 import 'package:tiomusic/src/rust/api/api.dart';
 import 'package:tiomusic/util/constants.dart';
 import 'package:tiomusic/widgets/input/number_input_and_slider_int.dart';
-import 'package:tiomusic/widgets/tap_to_tempo.dart';
+import 'package:tiomusic/widgets/input/tap_to_tempo.dart';
 
 class SetBPM extends StatefulWidget {
   const SetBPM({super.key});
